requirements.txt
flask
tensorflow
pillow
numpy
nltk
